+++ config/settings/test_demo_app.py
@@ -7,3 +7,5 @@ from .base import *  # noqa: F401,F403
 
 # noinspection PyUnresolvedReferences
 from .test import DATABASES, LOGGING  # noqa: F401
+
+DJANGO_STRUCTLOG_COMMAND_LOGGING_ENABLED = True