+++ django_structlog_demo_project/command_examples/apps.py
@@ -1,0 +1,6 @@
+from django.apps import AppConfig
+
+
+class CommandExamplesAppConfig(AppConfig):
+    name = "django_structlog_demo_project.command_examples"
+    default_auto_field = "django.db.models.AutoField"